+++ pyengine/Widgets/__init__.py
@@ -2,3 +2,4 @@ from pyengine.Widgets.Label import Label
 from pyengine.Widgets.Image import Image
 from pyengine.Widgets.Button import Button
 from pyengine.Widgets.Entry import Entry
+from pyengine.Widgets.Checkbox import Checkbox